+++ packages/caelus/src/index.ts
@@ -20,3 +20,4 @@ export * from "./profections.js";
 export * from "./firdaria.js";
 export * from "./releasing.js";
 export * from "./vedic.js";
+export * from "./directions.js";